+++ changelog.md
@@ -3,6 +3,13 @@
 ## 2.3.0
 * `NEW` `VSCode`: click the status bar icon to operate:
     * run workspace diagnostics
+* `NEW` `LuaDoc`: supports `[1]` as field:
+    ```lua
+    ---@class position
+    ---@field [1] number
+    ---@field [2] number
+    ---@field [3] number
+    ```
 * `FIX` loading workspace may hang
 * `FIX` `debug.getuservalue` and `debug.setuservalue` should not exist in `Lua 5.1`
 * `FIX` infer of `---@type class[][]`